Fix lint issues: explicit casting and explicit initialization

+++ examples/collision/square-platform.cpp
@@ -283,8 +283,8 @@ void Game::draw() {
 
 void start() {
     cout << ":: Game initialization!" << endl;
-    player.position.x = SCREEN_WIDTH / 2;
-    player.position.y = SCREEN_HEIGHT - player.size;
+    player.position.x = static_cast<float>(SCREEN_WIDTH) / 2;
+    player.position.y = static_cast<float>(SCREEN_HEIGHT) - player.size;
     player.update_rect();
 
     int block_size = 50;
